extracted alias logic to function
